separate config from auth application for loose couping unit test purpose
